--- gatsby-source-chec/gatsby/node/createSchemaCustomization.js
@@ -1,14 +1,0 @@
-const createSchemaCustomization = ({ actions: { createTypes } }) => {
-  createTypes(`
-    type ChecProduct implements Node {
-      categories: [ChecCategory] @link
-      images: [File] @link
-    }
-
-    type ChecCategory implements Node {
-      products: [ChecProduct] @link
-    }
-  `);
-};
-
-module.exports = createSchemaCustomization;
