+++ gatsby-source-chec/package.json
@@ -22,6 +22,6 @@
   },
   "dependencies": {
     "chec-request": "1.2.0",
-    "gatsby-source-filesystem": "2.11.1"
+    "gatsby-source-filesystem": "3.5.0"
   }
 }
