--- gatsby-source-chec/plugin-options.js
@@ -1,6 +1,0 @@
-const pluginOptions = (options) => ({
-  downloadImageAssets: false,
-  ...options,
-});
-
-module.exports = pluginOptions;
